[m133] Update infra to pick up recent changes from main branch
This is mostly to stop running on our Debian 10 VMs but includes a lot
of changes due to the lab move.

Bug: b/380037738
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/973616
Auto-Submit: Eric Boren <borenet@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>

+++ tools/gpu/gl/angle/GLTestContext_angle.cpp
@@ -476,13 +476,13 @@ void ANGLEGLContext::destroyGLContext() {
 
 void ANGLEGLContext::onPlatformMakeNotCurrent() const {
     if (!eglMakeCurrent(fDisplay, EGL_NO_SURFACE, EGL_NO_SURFACE, EGL_NO_CONTEXT)) {
-        SkDebugf("Could not reset the context 0x%x.\n", eglGetError());
+        SkDebugf("Could not reset the context 0x%x.\n", unsigned(eglGetError()));
     }
 }
 
 void ANGLEGLContext::onPlatformMakeCurrent() const {
     if (!eglMakeCurrent(fDisplay, fSurface, fSurface, fContext)) {
-        SkDebugf("Could not set the context 0x%x.\n", eglGetError());
+        SkDebugf("Could not set the context 0x%x.\n", unsigned(eglGetError()));
     }
 }
